remove the client::resource and server::resource modules
