+++ lightyear/src/lib.rs
@@ -127,9 +127,10 @@ pub mod prelude {
         pub use crate::client::prediction::plugin::{PredictionConfig, PredictionSet};
         pub use crate::client::prediction::predicted_history::{ComponentState, PredictionHistory};
         pub use crate::client::prediction::{Predicted, PredictionDespawnCommandsExt};
-        pub use crate::client::resource::Authentication;
         pub use crate::client::sync::SyncConfig;
-        pub use crate::connection::client::{ClientConnection, NetClient, NetConfig};
+        pub use crate::connection::client::{
+            Authentication, ClientConnection, NetClient, NetConfig,
+        };
         #[cfg(feature = "steam")]
         pub use crate::connection::steam::client::SteamConfig;
     }
