+++ lightyear/src/shared/replication/entity_map.rs
@@ -5,8 +5,13 @@ use bevy::prelude::{Entity, EntityWorldMut, World};
 use bevy::utils::hashbrown::hash_map::Entry;
 
 // TODO: another solution to avoid the orphan rule would be to implement MapEntities directly on the enum type?
-// Wrapper trait to avoid the orphan rule (so that I can implement MapEntities for external types)
+/// [`LightyearMapEntities`] is basically identical to [`MapEntities`]; I provide it to circumvent the orphan rule
+/// and to be able to implement [`MapEntities`] on external types.
 pub trait LightyearMapEntities {
+    /// Updates all [`Entity`] references stored inside using `entity_mapper`.
+    ///
+    /// Implementors should look up any and all [`Entity`] values stored within `self` and
+    /// update them to the mapped values via `entity_mapper`.
     fn map_entities<M: EntityMapper>(&mut self, entity_mapper: &mut M);
 }
